+++ src/variables/SEOMateVariable.php
@@ -10,6 +10,7 @@ namespace vaersaagod\seomate\variables;
 
 use craft\helpers\Template;
 use Twig\Markup;
+use vaersaagod\seomate\helpers\SEOMateHelper;
 use vaersaagod\seomate\SEOMate;
 
 /**
@@ -30,6 +31,10 @@ class SEOMateVariable
 
     public function breadcrumbSchema(array $breadcrumbArray): Markup
     {
+        $breadcrumbArray = array_map(static function (array $crumb) {
+            $crumb['url'] = SEOMateHelper::stripTokenParams($crumb['url']);
+            return $crumb;
+        }, $breadcrumbArray);
         $breadcrumbList = SEOMate::getInstance()->schema->breadcrumb($breadcrumbArray);
         return Template::raw($breadcrumbList->toScript());
     }